Moved Login and Register page components into page folder
